moved formatting.py into utils/
