+++ libs/langchain/langchain/prompts/base.py
@@ -5,10 +5,10 @@ import warnings
 from abc import ABC
 from typing import Any, Callable, Dict, List, Set
 
-from langchain.formatting import formatter
 from langchain.schema import BasePromptTemplate
 from langchain.schema.messages import BaseMessage, HumanMessage
 from langchain.schema.prompt import PromptValue
+from langchain.utils import formatter
 
 
 def jinja2_formatter(template: str, **kwargs: Any) -> str:
