+++ libs/langchain/langchain/utils/formatting.py
